+++ gameplay/src/PhysicsConstraint.h
@@ -12,7 +12,7 @@ namespace gameplay
 /**
  * Defines the base class for physics constraints.
  *
- * @see http://blackberry.github.io/GamePlay/docs/file-formats.html#wiki-Constraints
+ * @see http://gameplay3d.github.io/GamePlay/docs/file-formats.html#wiki-Constraints
  */
 class PhysicsConstraint
 {