+++ gameplay/src/PhysicsGenericConstraint.h
@@ -14,7 +14,7 @@ namespace gameplay
  * rigid bodies (or one rigid body and the world) where the
  * limits for all six degrees of freedom can be set individually.
  *
- * @see http://blackberry.github.io/GamePlay/docs/file-formats.html#wiki-Constraints
+ * @see http://gameplay3d.github.io/GamePlay/docs/file-formats.html#wiki-Constraints
  */
 class PhysicsGenericConstraint : public PhysicsConstraint
 {