+++ gameplay/src/lua/lua_Node.cpp
@@ -102,8 +102,8 @@ void luaRegister_Node()
         {"getWorldViewProjectionMatrix", lua_Node_getWorldViewProjectionMatrix},
         {"hasScriptListener", lua_Node_hasScriptListener},
         {"hasTag", lua_Node_hasTag},
-        {"isActive", lua_Node_isActive},
-        {"isActiveInHierarchy", lua_Node_isActiveInHierarchy},
+        {"isEnabled", lua_Node_isEnabled},
+        {"isEnabledInHierarchy", lua_Node_isEnabledInHierarchy},
         {"isStatic", lua_Node_isStatic},
         {"release", lua_Node_release},
         {"removeAllChildren", lua_Node_removeAllChildren},
@@ -120,12 +120,12 @@ void luaRegister_Node()
         {"scaleY", lua_Node_scaleY},
         {"scaleZ", lua_Node_scaleZ},
         {"set", lua_Node_set},
-        {"setActive", lua_Node_setActive},
         {"setAgent", lua_Node_setAgent},
         {"setAnimationPropertyValue", lua_Node_setAnimationPropertyValue},
         {"setAudioSource", lua_Node_setAudioSource},
         {"setCamera", lua_Node_setCamera},
         {"setCollisionObject", lua_Node_setCollisionObject},
+        {"setEnabled", lua_Node_setEnabled},
         {"setForm", lua_Node_setForm},
         {"setId", lua_Node_setId},
         {"setIdentity", lua_Node_setIdentity},
@@ -4002,7 +4002,7 @@ int lua_Node_hasTag(lua_State* state)
     return 0;
 }
 
-int lua_Node_isActive(lua_State* state)
+int lua_Node_isEnabled(lua_State* state)
 {
     // Get the number of parameters.
     int paramCount = lua_gettop(state);
@@ -4015,7 +4015,7 @@ int lua_Node_isActive(lua_State* state)
             if ((lua_type(state, 1) == LUA_TUSERDATA))
             {
                 Node* instance = getInstance(state);
-                bool result = instance->isActive();
+                bool result = instance->isEnabled();
 
                 // Push the return value onto the stack.
                 lua_pushboolean(state, result);
@@ -4023,7 +4023,7 @@ int lua_Node_isActive(lua_State* state)
                 return 1;
             }
 
-            lua_pushstring(state, "lua_Node_isActive - Failed to match the given parameters to a valid function signature.");
+            lua_pushstring(state, "lua_Node_isEnabled - Failed to match the given parameters to a valid function signature.");
             lua_error(state);
             break;
         }
@@ -4037,7 +4037,7 @@ int lua_Node_isActive(lua_State* state)
     return 0;
 }
 
-int lua_Node_isActiveInHierarchy(lua_State* state)
+int lua_Node_isEnabledInHierarchy(lua_State* state)
 {
     // Get the number of parameters.
     int paramCount = lua_gettop(state);
@@ -4050,7 +4050,7 @@ int lua_Node_isActiveInHierarchy(lua_State* state)
             if ((lua_type(state, 1) == LUA_TUSERDATA))
             {
                 Node* instance = getInstance(state);
-                bool result = instance->isActiveInHierarchy();
+                bool result = instance->isEnabledInHierarchy();
 
                 // Push the return value onto the stack.
                 lua_pushboolean(state, result);
@@ -4058,7 +4058,7 @@ int lua_Node_isActiveInHierarchy(lua_State* state)
                 return 1;
             }
 
-            lua_pushstring(state, "lua_Node_isActiveInHierarchy - Failed to match the given parameters to a valid function signature.");
+            lua_pushstring(state, "lua_Node_isEnabledInHierarchy - Failed to match the given parameters to a valid function signature.");
             lua_error(state);
             break;
         }
@@ -4917,42 +4917,6 @@ int lua_Node_set(lua_State* state)
     return 0;
 }
 
-int lua_Node_setActive(lua_State* state)
-{
-    // Get the number of parameters.
-    int paramCount = lua_gettop(state);
-
-    // Attempt to match the parameters to a valid binding.
-    switch (paramCount)
-    {
-        case 2:
-        {
-            if ((lua_type(state, 1) == LUA_TUSERDATA) &&
-                lua_type(state, 2) == LUA_TBOOLEAN)
-            {
-                // Get parameter 1 off the stack.
-                bool param1 = gameplay::ScriptUtil::luaCheckBool(state, 2);
-
-                Node* instance = getInstance(state);
-                instance->setActive(param1);
-                
-                return 0;
-            }
-
-            lua_pushstring(state, "lua_Node_setActive - Failed to match the given parameters to a valid function signature.");
-            lua_error(state);
-            break;
-        }
-        default:
-        {
-            lua_pushstring(state, "Invalid number of parameters (expected 2).");
-            lua_error(state);
-            break;
-        }
-    }
-    return 0;
-}
-
 int lua_Node_setAgent(lua_State* state)
 {
     // Get the number of parameters.
@@ -5451,6 +5415,42 @@ int lua_Node_setCollisionObject(lua_State* state)
         default:
         {
             lua_pushstring(state, "Invalid number of parameters (expected 2, 3, 4, 5 or 6).");
+            lua_error(state);
+            break;
+        }
+    }
+    return 0;
+}
+
+int lua_Node_setEnabled(lua_State* state)
+{
+    // Get the number of parameters.
+    int paramCount = lua_gettop(state);
+
+    // Attempt to match the parameters to a valid binding.
+    switch (paramCount)
+    {
+        case 2:
+        {
+            if ((lua_type(state, 1) == LUA_TUSERDATA) &&
+                lua_type(state, 2) == LUA_TBOOLEAN)
+            {
+                // Get parameter 1 off the stack.
+                bool param1 = gameplay::ScriptUtil::luaCheckBool(state, 2);
+
+                Node* instance = getInstance(state);
+                instance->setEnabled(param1);
+                
+                return 0;
+            }
+
+            lua_pushstring(state, "lua_Node_setEnabled - Failed to match the given parameters to a valid function signature.");
+            lua_error(state);
+            break;
+        }
+        default:
+        {
+            lua_pushstring(state, "Invalid number of parameters (expected 2).");
             lua_error(state);
             break;
         }